requirements.txt
absl-py==2.3.1
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
astunparse==1.6.3
certifi==2025.6.15
charset-normalizer==3.4.2
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
Cython==3.1.2
easydict==1.13
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.58.5
fsspec==2025.5.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.73.1
h5py==3.14.0
hf-xet==1.1.5
huggingface-hub==0.33.2
humanfriendly==10.0
idna==3.10
imageio==2.37.0
imbalanced-learn==0.13.0
insightface==0.7.3
Jinja2==3.1.6
joblib==1.5.1
keras-nightly==3.10.0.dev2025070803
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.3
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
namex==0.1.0
networkx==3.5
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.9
nvidia-nvtx-cu12==12.8.90
onnx==1.18.0
onnxruntime-gpu==1.22.0
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
pandas==2.3.1
pillow==11.2.1
prettytable==3.16.0
protobuf==6.31.1
psutil==7.0.0
py-cpuinfo==9.0.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytorch-triton==3.4.0+gitae848267
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
rich==14.0.0
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.16.0
setuptools==78.1.0
shapely==2.1.1
simsimd==6.5.0
six==1.17.0
sklearn-compat==0.1.3
stringzilla==3.12.5
sympy==1.14.0
tb-nightly==2.19.0a20250218
tensorboard-data-server==0.7.2
termcolor==3.1.0
tf_nightly==2.20.0.dev20250708
threadpoolctl==3.6.0
tifffile==2025.6.11
timm==1.0.16
torch==2.9.0.dev20250708+cu128
torchaudio==2.8.0.dev20250708+cu128
torchvision==0.24.0.dev20250708+cu128
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ultralytics==8.3.163
ultralytics-thop==2.0.14
urllib3==2.5.0
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
